app/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.6
django-cors-headers==4.3.0
django-otp==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
executing==2.0.1
google-api-python-client==1.6.4
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
ipdb==0.13.13
ipython==8.13.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jsonpickle==0.9.6
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
oauth2client==4.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
parso==0.8.3
phonenumbers==8.13.23
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==2.0.7
wcwidth==0.2.12
channels==4.0.0
celery==5.3.6
whitenoise==6.6.0
drf_spectacular==0.27.0
